修复bug,引入cglib 库高效的BeanCopier

+++ blog-application/src/main/java/com/dong/blog/application/BlogTypeApplication.java
@@ -1,7 +1,5 @@
 package com.dong.blog.application;
 
-import java.math.BigInteger;
-
 import org.dayatang.utils.Page;
 
 import com.dong.blog.application.dto.BlogTypeDTO;
@@ -10,6 +8,6 @@ public interface BlogTypeApplication extends BaseApplication<BlogTypeDTO, Long> 
 	
 	Page<BlogTypeDTO> getPage(BlogTypeDTO dto, int currentPage, int pageSize);
 	
-	BigInteger getTotal();
+	Long getTotal();
 
 }
